+++ src/context/TodoProvider.tsx
@@ -1,6 +1,6 @@
 import { ReactNode, createContext, useEffect, useState } from 'react';
-import { Task } from '../../types/todo';
-import { createTodo, deleteTodo, getTodos, updateTodo } from '../../apis/todoApi';
+import { Task } from '../types/todo';
+import { createTodo, deleteTodo, getTodos, updateTodo } from '../apis/todoApi';
 
 interface TodoDispatch {
   addTask: (content: string) => void;
